requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.7.0
Werkzeug==0.10.4
gunicorn==19.3.0
itsdangerous==0.24
wsgiref==0.1.2
